Move configuration into its own module
